fix(einsatz): route_avg_speed_kmh als DECIMAL fuer Hibernate-Validierung

+++ src/main/java/de/feuerwehr/manager/divera/AlarmdepechePdfService.java
@@ -78,7 +78,7 @@ public class AlarmdepechePdfService {
             return null;
         }
         int minutes = Math.max(1, (int) Math.round(alarm.getRouteDurationSec() / 60.0));
-        double speed = alarm.getRouteAvgSpeedKmh() != null ? alarm.getRouteAvgSpeedKmh() : 0;
+        double speed = alarm.getRouteAvgSpeedKmh() != null ? alarm.getRouteAvgSpeedKmh().doubleValue() : 0;
         return alarm.getRouteDistanceM()
                 + " m ("
                 + minutes
